harvester cleanup, metadata
