+++ harvester/harvester/parse/exceptions.py
@@ -24,19 +24,3 @@ class EmptyFileError(Exception):
     """
 
     pass
-
-
-class DataGenerationError(Exception):
-    """
-        Exception indicating the failure to generate some columns
-    """
-
-    pass
-
-
-class InsertError(Exception):
-    """
-        Exception indicating the failure to insert data
-    """
-
-    pass
