--- harvester/parse/database/cell_data/__init__.py
@@ -1,3 +1,0 @@
-from .cell_row import CellRow
-from .cell import Cell
-from .manufacturer_row import ManufacturerRow
